[jolla-camera] Reduce startup time. Contributes to JB#9890
Defer loading the camera roll until after the viewfinder is active,
or the user opens it, and then do the loading asynchronously.

Also take the opportunity to remove the control swapping ui which
contributes to compilation time and move the icon mapping to the
qml settings object instead of a js import.