[jolla-camera] Collect translations and qml types into single import. Contributes to JB#13627
Don't duplicate translations between the camera app and its settings
plugin.  And since things are moving around get rid of the settings
plugin and combine it with the qml types and translation loading from
the executable in a new com.jolla.camera plugin, so that there's no
depenency on the settings plugin from the executable.