[jolla-camera] Fix usability of pulley menu is gallery. Fixes JB#15633
If a nested flickable has a pressDelay it will filter press events
preventing the outer flickable from receiving them which meant flicks
were calculated relative to previous press and if this was sufficiently
far away the outer flickable would grab the mouse well in advance of the
inner flickable.  Removing the pressDelay from all but the outer most
flickable resolves this.