[jolla-camera] Verify file names don't exist before writing to them. Fixes JB#21632
The file name counter is persistent between runs of jolla-camera, and
concurrent instances so it will not overwrite files it created in
normal operation.  If another application writes to the same directory
with the same file naming convention or some unusual circumstance
leads to the counter being reset then it is possible to overwrite an
existing file, so add a simple check for existing files as an added
layer of protection.