[jolla-camera] Remove inotify cruft. Fixes JB#23090
Recheck the location state whenever we become active again.
In practice it does not have much over head and it's still a bit better
that using inotify directly which is really an overkill.

The best solution would be to get notified when the settings get changed but
dconf has been dropped from location.

Since it's unlikely under normal usage conditions that the setting will get
flipped while we are active then this approach is safe.