[camera] Sort camera reel by modification time. Fixes JB#34169
If external files with different naming scheme were copied to camera
directories, the order of the reel content got strange. Taking a
picture and swiping to reel showed the correct image, but user
was also taken into the middle of the reel.

Before commit 2f5322c56fe dateTaken was used, but reportedly that doesn't
work with root type file, and doesn't seem available for videos either.