[jolla-camera] Initialise PositionSource.active in a safe manner. Contributes to JB#38358
The QDeclarativePositionSource.active property is not implemented in
a well-behaved manner (in that it actually exposes the state of an
asynchronous/queued operation, and the value will be reset to false
when the internal position source is initialised (which only occurs
upon component completion in the default case)).

This commit ensures that we work around this issue by initialising
the property only after the component completes, and thus after the
default internal position source has been constructed appropriately.

Contributes to JB#38358