[camera] Add remaining time for recording, with auto-shutoff. Fixes JB#42069. Fixes JB#46400
Remaining video recording time is calculated using the video and audio
bitrates + 5% overhead. If the current storage is a vfat partition, this
can't be more than the 4GB filesize limit. If device storage is being
used, 100MB is reserved to avoid filling it.

When recording is started, the remain time counts down. When this reaches
zero, recording is automatically stopped. If current remain time is zero,
the start recording button is disabled.

The remain time is shown before recording is started in hours and minutes,
and as a minute countdown during the last ten minutes.