[camera] Allow whitebalance and exposure level on all the scene modes. Contributes to JB#53786
The whitebalance does make some sense on the modes, especially on HDR,
though to be seen if we keep it for the special modes on the next release
and how it really works on different devices. Exposure slider allowed
for now as it's paired on the UI and doesn't hurt getting some
experience also on that.