[jolla-camera] Read the capture directories to populate the camera roll. JB#54283 OMP#JOLLA-144
Images and Videos are in separate graphs in Tracker 3, so to get a
combined view with QtDocGallery two different models have to be spliced
together which is acheivable but complicated. The easier, more robust
option is to just read from file system directly and sort on the file
names which encode the capture time. Because of the limited scope this
will be as fast or faster than querying tracker.